child category feature- show category

+++ app/Http/Controllers/Backend/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Backend;
 use App\DataTables\CategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use Str;
 
@@ -99,6 +100,10 @@ class CategoryController extends Controller
     public function destroy(string $id)
     {
         $category = Category::find($id);
+        $subCategory = SubCategory::where('category_id', $category->id)->count();
+        if($subCategory > 0) {
+            return response(['status' => 'error', 'message' => 'This category has sub categories. Please delete them first!']);
+        }
         $category->delete();
 
         return response(['status' => 'success', 'message' => 'Deledted successfully!']);
